Distribute handlers
