+++ internal/handlers/insert_data_handler.go
@@ -1,13 +1,11 @@
 package handlers
 
 import (
-	"encoding/csv"
 	"encoding/json"
 	"fmt"
 	"io"
 	"log"
 	"net/http"
-	"os"
 	"strings"
 	"sync"
 	"time"
@@ -45,6 +43,14 @@ func fetchJSON(url string, target interface{}) error {
 }
 
 // InsertDataHandler обрабатывает вставку данных в базу данных
+// @Summary Вставка данных
+// @Description Вставляет данные расписания и экзаменов в базу данных
+// @Tags InsertData
+// @Accept json
+// @Produce json
+// @Success 200 {object} map[string]string "message: Data inserted successfully"
+// @Failure 500 {object} map[string]interface{} "errors: [error messages]"
+// @Router /api/v1/insert-data [post]
 func (a *App) InsertDataHandler(c echo.Context) error {
 	structureURL := "https://lks.bmstu.ru/lks-back/api/v1/structure"
 	var structure models.Structure
@@ -168,81 +174,4 @@ func appendError(mu *sync.Mutex, errors *[]string, message string) {
 	mu.Lock()
 	defer mu.Unlock()
 	*errors = append(*errors, message)
-}
-
-// WriteScheduleToFile сохраняет расписание в CSV файл
-func (a *App) WriteScheduleToFile(c echo.Context) error {
-	var scheduleItems []models.ScheduleItem
-
-	// Загрузка данных с подгрузкой аудиторий и преподавателей
-	if err := a.DB.Preload("Teachers").Preload("Audiences").Find(&scheduleItems).Error; err != nil {
-		return c.JSON(http.StatusInternalServerError, map[string]string{"error": "Failed to fetch schedule items"})
-	}
-
-	filePath := "/usr/src/semesterly/data/schedule.csv"
-	if err := writeToCSV(filePath, scheduleItems); err != nil {
-		return c.JSON(http.StatusInternalServerError, map[string]string{"error": "Failed to write schedule to file"})
-	}
-
-	return c.JSON(http.StatusOK, map[string]string{"message": "Schedule written to file successfully"})
-}
-
-// writeToCSV записывает данные расписания в CSV файл
-func writeToCSV(filePath string, scheduleItems []models.ScheduleItem) error {
-	file, err := os.Create(filePath)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	writer := csv.NewWriter(file)
-	defer writer.Flush()
-
-	// Заголовки CSV
-	writer.Write([]string{"Day", "Time", "Week", "Stream", "EndTime", "StartTime", "Discipline", "Permission", "Teachers", "Audiences"})
-
-	for _, item := range scheduleItems {
-		// Форматирование списка преподавателей
-		teachers := make([]string, len(item.Teachers))
-		for i, teacher := range item.Teachers {
-			teachers[i] = fmt.Sprintf("%s %s %s", teacher.LastName, teacher.FirstName, teacher.MiddleName)
-		}
-
-		// Форматирование списка аудиторий
-		audiences := make([]string, len(item.Audiences))
-		for i, audience := range item.Audiences {
-			audiences[i] = audience.Name
-		}
-
-		// Запись строки в CSV
-		err := writer.Write([]string{
-			fmt.Sprintf("%d", item.Day),
-			fmt.Sprintf("%d", item.Time),
-			item.Week,
-			item.Stream,
-			item.EndTime,
-			item.StartTime,
-			item.Discipline.FullName,
-			item.Permission,
-			strings.Join(teachers, "; "),
-			strings.Join(audiences, "; "),
-		})
-		if err != nil {
-			return fmt.Errorf("failed to write to CSV: %w", err)
-		}
-	}
-
-	return nil
-}
-
-// GetData отправляет JSON со всем расписанием из базы данных
-func (a *App) GetDataHandler(c echo.Context) error {
-	var scheduleItems []models.ScheduleItem
-
-	// Загрузка данных с подгрузкой аудиторий и преподавателей
-	if err := a.DB.Preload("Teachers").Preload("Audiences").Find(&scheduleItems).Error; err != nil {
-		return c.JSON(http.StatusInternalServerError, map[string]string{"error": "Failed to fetch schedule items"})
-	}
-
-	return c.JSON(http.StatusOK, scheduleItems)
 }
